Tutorial 1-Commit 3 Activating Models
In the activating models section, I had to edit my mysite/settings.py file
to tell the project that the polls app is installed.I did this by
changing the INSTALLED_APPS in the setting to include the string'polls,'.
After this I ran python manage.py makemigrations polls to tell Django that
I made changes to my models(made new ones in this part) and I want to
store the changes as a migration.I ran the sqlmigrate polls command that
takes migration and returns their SQL. Lastly, I  ran migrate again to
create model tables in my database.