Tutorial 2 Commit 2: Adding Choice to polls/admin
After registering Choice, I removed the register() call for the Choice model and edited the
Question code. This was to let Django know that Choice objects
are edited on the Question admin page, and to provide 3 default fields
for choices.Because the fields take up a lot of space,
I edited the polls/admin file again to change my ChoiceInLine to be
Tabular.. instead of Stacked.. so objects would be displayed in a more
compact,table based format.