Tutorial 2 Commit 4: Customizing the admin
I customized the admin look by creating a templates directory in my project directory
and added the DIRS option to TEMPLATES in mysite/settings.py.DIRS is a search path , a list
of file system directories to check when loading Django templates.
Then,I created an admin directory inside templates and copied a template
from the default admin template to the customized template.