Tutorial 3 Commit 2 : Updating the index view
I continued to write more views- where each view is responsible for
either retuning an HttpResponse object containing the content of the
requested page or raising an exception such as Http404.

I added an index() view that displays 5 poll questions according to
publication date.I was directed to  change the way the page looks by using
Django's template system to seperate the design by creating a template for
view to use.

I called a directory called templates in polls so Django can
look for templates in there.And within templates, I created a polls
directory with an index.html file and updated it with the given code.
After this, I had to update the index view in polls/views to use the
template.I checked my browser and saw a bulleted list containing the
"What's up" question.