Tutorial 3 Commit 3: render() and 404 error
I continued to edit my polls/views.py file by using the render() shortcut
template and updating the index() view using render(). The render()
function takes the request object as its first argument, a template name
as its second,and a dictionary as its third. It returns an HttpResponse
object of the given template.

I went on to the raising a 404 error section where I edited the question
detail view.If a question with the requested ID does not exist, the view
raises a Http404 exception.I had to make a new detail.html template
containing { {Question} }.