Tutorial 4 Commit 1: Writing an HTML form
I updated the poll detail template with code that contained an HTML form
element and my polls/views.py file with a vote() function.In order to
redirect the results page when someone votes, I changed my results view
with the given code. My next step was to create a results.html template
so that the results page gets updated with each vote.
I voted in my question and was able to see an updated results page.