Tutorial 5 Commit 1 : Identifying/exposing the bug
The objective was to create automated tests that check the operation of
the code. Automated tests are done by the system, saving manual testing
time. Creating tests is useful because they save time,identify and prevent
problems,make code more attractive,and foster better collaboration.

To write my first test, I had to identify a bug by using shell.When I
created a question that lies in the future,the Question  change list
claims it was published recently. This is wrong, we have identified a bug.

The next step was to create a test to expose the bug. In polls/tests.py I
wrote the given code that  creates a Question instance with a publication
date in the future.

Then,I ran my test using python manage.py test polls that looked for tests
in the polls application. The test result was FAILED, with an
AssertionError: True ! = False meaning that the question with a future
publication date was identified as being recently published. The test failed.