Tutorial 5 Commit 2 : Fixing the bug
Because the bug lies in the Question.was_published_recently()method, I
amended the method in polls.models.py so that it only returns "True" if
the date is in the past.

I ran the test and got OK,meaning that the bug was corrected and the test
passed.To make the method more reliable, I added two more test methods
that confirm correct results for past,recent,and future questions.