Tutorial 5 Commit 3:Test client and imporving view
I used shell to set up the test environment and imported the test client
class on it. The client class created a new question, "Who is your
favorite Beatle" and saved it.

Also, I was directed to improve the get_queryset() method from my
IndexView class in polls/views.py.To do this, I imported timezone
and amended the get_queryset method with the given code.Now the
changed method can check the date by compaing it with timezone.now().