Tutorial 5 Commit 4:Testing new and Detail views
I modified my polls/tests file by creating a shortcut function
that creates questions and added a new test class.

In this modification/new class:
-create_question helps not be repetitive when creating new questions

-test_index_view_with_no_questions verifies that the latest question list
is empty and checks the message "No polls are avaliable".

-using test_index_view_with_a_past_question,a question is created and
verified it is on the list

-using test_index_view_with_a_future_question, we create a question
with a future publication date

I imporved my DetailView by adding a constraint so users cannot reach
future questions and added the class QuestionIndexDetailTests in
polls/tests to make sure that questions with past publication dates are
displayed and ones with future dates are not.