Tutorial 6 Commit 2: Background image
To add a background image on my server, I created an images subdirectory
in the polls folder from the static directory and put image called
background.gif. Lastly, I updated my stylesheet with given code to add the
image to the bottom right of my page.
I reloaded the server and saw green links and the image on the bottom
right.